requirements.txt
Flask
Flask-Cors
flask-jwt-extended
mysql-connector-python
python-dotenv
marshmallow